requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
pillow==10.1.0
google-generativeai==0.3.2
requests==2.31.0
plotly==5.18.0
python-dotenv==1.0.0
matplotlib==3.8.2
scikit-learn==1.3.2 